feat: move schemas folder to out of app folder
