+++ src/app/actions/todos/add-todo-action.ts
@@ -2,8 +2,8 @@
 
 import { revalidatePath } from 'next/cache'
 
+import { TTodoSchema } from '@/schemas/todoSchema'
 import { createTodo } from '@/app/services/api-todos'
-import { TTodoSchema } from '@/app/schemas/todoSchema'
 import { getErrorMessage } from '@/app/utils/getErrorMessage'
 
 export const addTodoServerAction = async (formData: TTodoSchema) => {
